class 03 - server: authentication flow using GitHub and the server is now creating JWT Token, to check if the user is logged in and keep him logged in; web: Splitting the Home page into components and the 'SignIn' link now directs to a url that generates an authentication request code;

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify(); /* se não tiver um token, o restante do código pára */
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     /* esse parâmetro 'request' do fastify possui o id */
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -29,6 +36,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
     return memory;
   });
 
@@ -44,13 +54,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "77059a5c-9ae1-4ce0-bd7e-e9a58924639e",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -61,7 +71,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false),
     });
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
-    const memory = await prisma.memory.update({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -71,19 +90,29 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
       },
     });
+
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
-    const memory = await prisma.memory.delete({
+
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
-    return memory;
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    });
   });
 }
